+++ app/src-tauri/src/app.rs
@@ -1,0 +1,312 @@
+use ashuk_core::{
+    compresser::{
+        compress_to_target_extension, CompressError, CompressOptions, Result as CompressResult,
+        Status,
+    },
+    format_meta::{CompressOptionsContext, ImageFormat, ProcessStrategy},
+};
+use rayon::prelude::*;
+use serde::{Deserialize, Serialize};
+use tauri::Manager;
+
+use std::collections::HashMap;
+use std::sync::Mutex;
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct InputResult {
+    pub path: String,
+    pub size: u64,
+    pub extension: String,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct FileContext {
+    pub status: Status,
+    pub input: InputResult,
+    pub output: Option<CompressResult>,
+}
+
+type FileList = HashMap<String, FileContext>;
+
+pub struct FileState {
+    files: Mutex<FileList>,
+    options: Mutex<CompressOptions>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub enum EmitFileOperation {
+    Create,
+    Update,
+    Compress,
+    Clear,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct EmitFileRequestBody {
+    files: Option<FileList>,
+    operation: EmitFileOperation,
+    options: Option<CompressOptions>,
+}
+
+impl FileState {
+    pub fn new(files: FileList) -> Self {
+        // set default option
+        let option = get_compress_options_context().unwrap();
+        Self {
+            files: Mutex::new(files),
+            options: Mutex::new(CompressOptions {
+                extension: option[0].extension.clone(),
+                quality: Some(option[0].default_value),
+            }),
+        }
+    }
+
+    pub fn add_file(&self, file_path: &str) -> FileContext {
+        let mut files = self.files.lock().unwrap();
+        // init data
+        let file = FileContext {
+            status: Status::Initialized,
+            input: InputResult {
+                path: file_path.to_string().clone(),
+                size: std::fs::metadata(&file_path)
+                    .expect("There is no file.")
+                    .len(),
+                extension: ImageFormat::from_path(&file_path)
+                    .unwrap()
+                    .get_representative_ext_str(),
+            },
+            output: None,
+        };
+        // update hashmap
+        files.entry(file_path.to_string()).or_insert(file.clone());
+
+        file
+    }
+
+    pub fn update_file(&self, path: String, file_status: FileContext) -> FileContext {
+        let mut files = self.files.lock().unwrap();
+        let new_file = FileContext {
+            status: file_status.status,
+            input: file_status.input,
+            output: file_status.output,
+        };
+        // update hashmap
+        files.entry(path).or_insert(new_file.clone());
+
+        new_file
+    }
+
+    pub fn get_files(&self) -> FileList {
+        let files = self.files.lock().unwrap();
+        files.clone()
+    }
+
+    pub fn clear(&self) {
+        let mut files = self.files.lock().unwrap();
+        files.clear();
+    }
+
+    pub fn update_options(&self, _options: Option<CompressOptions>) {
+        let mut options = self.options.lock().unwrap();
+        *options = _options.unwrap();
+    }
+
+    pub fn process_compress(
+        &self,
+        app_handle: &tauri::AppHandle,
+        emitter_name: &str,
+        path: std::string::String,
+        file: FileContext,
+    ) {
+        // compress option
+        let options = self.options.lock().unwrap();
+        // update state
+        let updated_file = self.update_file(
+            path.clone(),
+            FileContext {
+                status: Status::Pending,
+                input: file.clone().input,
+                output: None,
+            },
+        );
+        // notify to client for start compressing
+        notify_file_to_client(&app_handle, &updated_file, emitter_name);
+
+        // compress image
+        let result = compress_to_target_extension(
+            &path,
+            CompressOptions {
+                extension: options.extension.clone(),
+                quality: options.quality,
+            },
+        );
+
+        if result.is_ok() {
+            // update state
+            let updated_file = self.update_file(
+                path,
+                FileContext {
+                    status: Status::Success,
+                    input: file.input,
+                    output: Some(result.unwrap()),
+                },
+            );
+            // notify to client for success
+            notify_file_to_client(&app_handle, &updated_file, emitter_name);
+        } else {
+            // handle error
+            let status = match &result.err().unwrap() {
+                CompressError::Unsupported(..) => Status::Unsupported,
+                _ => Status::Failed,
+            };
+            // update state
+            let updated_file = self.update_file(
+                path,
+                FileContext {
+                    status: status,
+                    input: file.input,
+                    output: None,
+                },
+            );
+            // notify to client for failure
+            notify_file_to_client(&app_handle, &updated_file, emitter_name);
+        }
+    }
+}
+
+fn notify_file_to_client<T: ?Sized + Serialize>(
+    app_handle: &tauri::AppHandle,
+    file_context: &T,
+    emitter_name: &str,
+) {
+    let serialized = serde_json::to_string(&file_context).expect("Invalid data format");
+    app_handle.emit_all(&emitter_name, serialized).unwrap();
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+struct SupportedFormatMeta {
+    ext: String,
+}
+
+#[tauri::command]
+fn get_compress_options_context() -> Result<Vec<CompressOptionsContext>, String> {
+    let options = ImageFormat::get_formats()
+        .iter()
+        .map(|x| x.get_compress_options_context())
+        .collect::<Vec<CompressOptionsContext>>();
+    Ok(options)
+}
+
+#[tauri::command]
+fn get_supported_extensions() -> Result<Vec<SupportedFormatMeta>, String> {
+    let extensions = ImageFormat::get_formats()
+        .iter()
+        .map(|x| {
+            x.extensions_str().iter().map(move |y| SupportedFormatMeta {
+                ext: <str as ToString>::to_string(*y),
+            })
+        })
+        .flatten()
+        .collect::<Vec<SupportedFormatMeta>>();
+    Ok(extensions)
+}
+
+fn compress_file_handler(app: &tauri::AppHandle) {
+    let emitter_name = "listen-file";
+    let file_state = FileState::new(HashMap::new());
+    let app_handle = app.app_handle();
+    // listen file input
+    let emit_file_create = app_handle
+        .app_handle()
+        .listen_global("emit-file", move |event| {
+            // get task
+            let task: EmitFileRequestBody = serde_json::from_str(&event.payload().unwrap())
+                .expect("JSON was not well-formatted");
+
+            match task.operation {
+                // notify to client for ready
+                EmitFileOperation::Create => {
+                    // check
+                    if let Some(v) = &task.files {
+                        // process parallelly
+                        v.into_par_iter().for_each(|(path, _)| {
+                            let add_file = file_state.add_file(&path);
+                            notify_file_to_client(&app_handle, &add_file, emitter_name);
+                        });
+                    };
+                }
+                // compress
+                EmitFileOperation::Compress => {
+                    // check
+                    if let Some(v) = &task.files {
+
+                        // wheather process parallelly or not
+                        let paralell_processable =
+                            v.into_par_iter().all(|(path, _)|
+                            // process strategy with multithreading depends on it's library implementation
+                            matches!(
+                                ImageFormat::process_strategy(&ImageFormat::from_path(&path).unwrap()),
+                                ProcessStrategy::Parallel
+                            ));
+
+                        match paralell_processable {
+                            true => {
+                                // process parallelly
+                                v.clone()
+                                    .into_par_iter()
+                                    // skip compressed file
+                                    .filter(|(_, file)| !matches!(file.status, Status::Success))
+                                    .for_each(|(path, file)| {
+                                        file_state.process_compress(
+                                            &app_handle,
+                                            &emitter_name,
+                                            path,
+                                            file,
+                                        )
+                                    });
+                            }
+                            _ => {
+                                // process seriesly
+                                v.clone()
+                                    .into_iter()
+                                    // skip compressed file
+                                    .filter(|(_, file)| !matches!(file.status, Status::Success))
+                                    .for_each(|(path, file)| {
+                                        file_state.process_compress(
+                                            &app_handle,
+                                            &emitter_name,
+                                            path,
+                                            file,
+                                        )
+                                    });
+                            }
+                        };
+                    };
+                }
+                EmitFileOperation::Update => {
+                    file_state.update_options(task.options);
+                }
+                EmitFileOperation::Clear => {
+                    file_state.clear();
+                    notify_file_to_client(&app_handle, &EmitFileOperation::Clear, "listen-clear-file");
+                }
+            }
+        });
+}
+
+pub fn init_app() {
+    tauri::Builder::default()
+        .invoke_handler(tauri::generate_handler![
+            get_supported_extensions,
+            get_compress_options_context,
+        ])
+        .setup(|app| {
+            compress_file_handler(&app.app_handle());
+
+            #[cfg(debug_assertions)]
+            app.get_window("main").unwrap().open_devtools();
+            Ok(())
+        })
+        .run(tauri::generate_context!())
+        .expect("error while running ashuk");
+}
