+++ app/src-tauri/src/logger.rs
@@ -1,0 +1,23 @@
+use std::env;
+use std::io::Write;
+
+use env_logger;
+
+pub fn init_logger() {
+    env::set_var("RUST_LOG", "debug");
+    env_logger::Builder::from_default_env()
+        .format(|buf, record| {
+            let ts = buf.timestamp();
+            writeln!(
+                buf,
+                "[{} {} {}] {} {}:{}",
+                ts,
+                record.level(),
+                record.target(),
+                record.args(),
+                record.file().unwrap_or("unknown"),
+                record.line().unwrap_or(0),
+            )
+        })
+        .init();
+}
